fix(core):h_DWORD适配ctypes增加描述

+++ hkws/core/type_map.py
@@ -8,7 +8,7 @@ h_INT = c_int
 h_WORD = c_uint16
 h_LONG = c_long
 h_FLOAT = c_float
-h_DWORD = c_uint32
+h_DWORD = c_ulong  # 64bit:c_ulong    32bit:c_uint32
 
 h_VOID_P = c_void_p
 h_HWND = c_void_p  # handle of window
